+++ NEWS.md
@@ -4,6 +4,7 @@
 * Added unit tests for SCD_effect_sizes shiny app.
 * Added intervention_phase argument to NAP(), SMD(), and other user-facing effect size calculation functions. 
 * Added warn argument to Tau_BC() function, to control printing of messages regarding baseline trend test.
+* Corrected a cosmetic bug in a warning message of batch_calc_ES() when scale argument is specified as a variable and ES includes "LOR".
 
 # SingleCaseES 0.5.0