+++ NEWS.md
@@ -4,6 +4,7 @@
 * Added unit tests for SCD_effect_sizes shiny app.
 * Added intervention_phase argument to NAP(), SMD(), and other user-facing effect size calculation functions. 
 * Added warn argument to Tau_BC() function, to control printing of messages regarding baseline trend test.
+* Added Kendall's rank correlation to Tau_BC() function.
 
 # SingleCaseES 0.5.0