+++ NEWS.md
@@ -5,6 +5,8 @@
 * Added intervention_phase argument to NAP(), SMD(), and other user-facing effect size calculation functions. 
 * Added warn argument to Tau_BC() function, to control printing of messages regarding baseline trend test.
 * Added Kendall's rank correlation to Tau_BC() function.
+* Corrected a cosmetic bug in a warning message of batch_calc_ES() when scale argument is specified as a variable and ES includes "LOR".
+
 
 # SingleCaseES 0.5.0