+++ NEWS.md
@@ -1,3 +1,5 @@
+# SingleCaseES 0.6.0.9999
+
 # SingleCaseES 0.6.0
 
 * Changes to batch_calc_ES()